Add descriptor field to Feature

+++ include/Utils/Frame.hpp
@@ -20,7 +20,7 @@ class Frame {
     Frame(cv::Mat image,
         std::vector<Feature> features,
         Pose pose,
-        uint64_t frame_id,
+        int frame_id,
         bool is_keyframe,
         bool is_processed);
 
@@ -40,7 +40,7 @@ class Frame {
     void set_image(cv::Mat image);
     void set_features(std::vector<Feature> features);
     void set_pose(Pose pose);
-    void set_frame_id(uint64_t id);
+    void set_frame_id(int id);
 
     // ----
     // Data
@@ -73,7 +73,7 @@ class Frame {
     bool is_processed_;
 
     // 
-    uint64_t frame_id_;
+    int frame_id_;
 };
 
 #endif
